Couple of updates
This refactor slightly changes the location of a few files, specifically the `keymap.c`
file which is used to determine the maps for the keyboard.  This is due to an upcoming
change to the project to allow support for more keyboards*

Also updated the default keymap, changing the Numpad Asterisk key to be Print Screen when
the Fn modifier is pressed.  We also change the default behaviour for SysReq to be the
Pause key mapping.

* Multiple Keyboard Support.  I recently aquired an old 1988 Square Badge Model M (1390401)
and have taken the decision to work on an external convertor which will support the same
code base, so this firmware code will extend beyond that of just the Model F.