Performance updates
Change how we send commands to the keyboard controller from `pm_sm_put_blocking` to `pm_sm_put`, and reduce debugging output during LED toggling.

This is to try and reduce the likelyhood of the keyboard locking up when spamming any of the Lock keys.  Unfortunately it doesn't completely resolve the issue, but does seem to reduce the chance of it happening.

To be honest though, who spams a Lock key on a keyboard? Behave!