Update Keyboard Init Process
There has always been an issue where upon power up, the process of initialisation would always
result in an initial invalid read from the keyboard (resulting in getting 0x54 with invalid
parity bit) instead of 0xAA which signifies the keyboard has passed BAT.  Originally we would
mitigate against this by passing 0xAA to the event processor function after a 100ms delay, so
tricking the initialisation process to begin.  This would then cause the converter to send
multiple 0xFF reset commands to the keyboard on power up.

Although this worked, it was messy.

As part of the Keyboards initialisation process, it will pull the CLOCK line low, and only bring
it high once it has passed its own self test prior to sending 0xAA (or the results of BAT).  To
overcome the problem of receiving this as an invalid stream, we now perform the following:

On SM Initialisation:
Set DATA High
Set CLOCK Low

Update PIO Routing to add an aditional init step to wait for CLOCK to go HIGH.

This now ensures we recieve the correct signal on RX during initial power up (0xAA on BAT)

We've also updated the pcat_lock_leds_handler to become pcat_task which now handles the main
initialisation process to begin with (but does still handle LED state changes).  Now, we have
a timer which checks for CLOCK to be HIGH, and upon that condition, if the keyboard is still
in an unitialised state, after 3 seconds it will issue 0xFF to ask the Keyboard to reset.

When initially powering on, as we recieve 0xAA correctly now, this function is automatically
bypassed due to conditions, however, if we are flashing the firmware or simply resetting the
RP2040 which connected, it will now perform a reset after 3 seconds of CLOCK being HIGH.

I may still update/change this further, but my intention from this going forward, is to start
supporting other IBM AT-protocal keyboards (I have an '88 1390131 to test with now), and also
looking at a new revision of the converter hardware which will be more of an in-line adaptor
allowing multiple keyboards to be supported.