Multiple updates
Quite a few chunks have been changes as part of this commit, but a quick summary of what has been done:

- Rework how we determine the Clock Multiplier for the State Machine.  This now allows paramater-based
adjustment by setting the desired clock speed in kHz.
- Fix issue with sending data to AT/PS2 Keyboard.  IBM Model F and '88 Model M seems to be quite
forgiving in when it receives commands from the host, however when testing with other PS2 keyboards,
the correct flow & checking methods were quite critical to ensure the SM didn't lock up.
- Add new keyboard layout for IBM Model M (1390131) as I now have this keyboard to develop and test
with.
- Update PCAT flow to allow extended key code events (E0 and E1)
- For now, we are updating the keymap lookup to a 16x16 array to quickly take in the E0 and E1 extended
codes.
- Rework how we initialise the keyboard to use a timer instead of an arbituary delay/command being sent
to trigger the flow.  This also fixes the 0x54 initialisation bug.
- Correctly reset and restart the State Machine given certain conditions (Restart or Data Error)
- Change HID Reporting Interval from 512ms to 128ms (bcdInterval reduced from 10 to 8)