Start of Multi-Keyboard and Protocol Support
This commit is the start of future enhancements to allow multiple Keyboard and Protocol support.

Previously, source files would need to be modified to allow the use of the extra Model M Keyboard, so this has now been changed to allow selection at build time, and no longer require updating of source files to change/select different keyboards.

A quick overview of changes:

- Better support for Multiple Keyboards (as added)
- Add support for Compile-time Protocol selection (based on `keyboard.config` for the relevant Keyboard we are building for)
- Add more output to the start of the program to indicate build time & compile time options relating to this change
- Restructure Source files to allow better compilation/modularisation of the above changes.
- Update documentation around Keyboards
- Rename project from `Model F 5170 Converter` to `RP2040 Keyboard Converter` to encompass the above changes
- Update entrypoint.sh to use `cmake` to build instead of `make`, and ensure we run a full cleanup prior to building