Update IRQ event and handling. Instead of triggering an IRQ event within the PIO State Machine itself, we are going to change the initiator from `pis_interrupt0` to `pis_sm0_rx_fifo_not_empty` to to trigger events automatically when the RX FIFO has data in it.
We also don't want to call `pio_interrupt_clear` from within the function triggered from the IRQ event.  This has the (unwanted) behaviour of clearing interrupts for any queued data within the RX FIFO, so would then causes key presses to be out of sync.

This has the benefit of also eliminating specific race conditions for when the input stream becomes congested due to either excessive debugging output to UART, or from lag in any other part of the keyboard handling process.