Initial support for Scancode Set 3 (AT Protocol) This change is to add support for the AT Protocol Scancode Set 3.  This is used primarily on Terminal keyboards which use the AT Protocol.
We add 3 new steps to the Keyboard State:

- INIT_READ_ID_1
- INIT_READ_ID_2
- INIT_SETUP

Only if we are buiding against a Keyboard which utilises Set 3 do these steps take effect.  Directly following BAT, Terminal keyboards (which utilise Set 3) send 2 extra bits of information which relate to the ID of this Keyboard.  In some cases, this is configurable on the Keyboard itself by using a DIP switch to select/change each bit.  This allows each ID Field to range from 0x80 to 0xFF.

- INIT_READ_ID_1
This directly reads the first 8 bytes following successful BAT, and updates the `keyboard_id` variable with this value
For example, ID_1 returns as 0x80, so `keyboard_id` becomes 0x80FF

- INIT_READ_ID_2
This reads the second stream of 8 bytes following BAT, and updates `keyboard_id` with this part.
For example, ID_2 returns as 0x8F, so now `keyboard_id` becomes 0x808F

- INIT_SETUP
Once both ID bits have been read, this step then ensures that the Keyboards keys are all set to Make/Break.  By default, Terminal keyboards are set to Make codes only, so will only ever present codes when a key is pressed, with no break code sent on release.  Here we ensure that the keyboard has accepted this configuration option once we send 0xF8 to the Keyboard.

The keyboard process then flows through the rest of the initialisation process as-normal.

We've also updated the mechanism used within `keyboard_interface_task` which is called on loop from `main.c`.  In here, we now handle (better) multiple timeout events based on Keyboard State:
- Reduced polling interval to 500ms from 1000ms (function is called every loop, but only processes every 500ms internally to itself)
- Timeout any of the INIT_READ_ID_* steps and INIT_SETUP after 500ms if the entire journey doesn't complete within that time.  This step really should only take around 20ms MAX to complete.
- Added a timeout to SET_LOCK_LEDS to time out after 500ms if the state remains.