Fix PIO Assignment for AT/PS2 Keyboards following LED Code.
When sending commands to the Keyboard, we hard-set pio1 as the receiver, however, with the new dynamic PIO assignment code, this changed.

This didn't come to light until testing the MicroSwitch 122ST13 keyboard, as using the Model M and Model F boards, these default to the correct typematic settings, so didn't directly show this issue.