Add Mouse Support
This adds support for Mice, as well as overhauls key aspects of how PIO and StateMachines are managed.

As part of this, PS2 mouse support is now available, and can be enabled at compile time by specifying the following option during build:

`docker-compose run -e KEYBOARD="modelf/pcat" -e MOUSE="at-ps2" builder`

The build process will allow either a Keyboard, Mouse or both to be specified.

PS2 support for Mice so far is still WIP.  To build & test this functionality, the only PS2 mouse which I have is a standard 2-button one, so for
 supporting Scroll Wheel mice, as well as 5 button mice, has been developed purely from reading various documentation on how PS2 mice operate.  A
s such, the code relating to these types of mice will most likely change in future as I gather more devices to test with.

PIO Code for the PS2 mouse uses the same source as that of the Keyboard.  As such, build process now looks for either `keyboard_interface.pio`, `
mouse_interface.pio` or (in the case of the AT/PS2 protocol), a generic `interface.pio` file is used.  As we then define and load the code into m
emory, we now determine which PIO and StateMachine instance will be used, so avoiding any need for unnecessary configuration or fudging needed to be able to seamlessly build against various devices.

Unlike the Keyboard interface, the Mouse one does not use a buffer for movement & button data.  As such, we drive all responses and reports directly from an interrupt event triggered as we receive data from the mouse.  The mouse is also configured to operate in its default `report` mode, so response time and latency is kept as minimal as possible.

USB Configuration has been updated to disable standby support for the device, as we're driving Keyboard and Mouse devices directly from VCC (5v) from the USB host, the majority of devices which could be attached won't support any form of standby.  As such we have also removed the relevant standby and wakeup calls from the main `keyboard_interface_task` loops, which again has had a noticeable change in reducing latency.

We've also renamed a few things:

`at` protocol has now been renamed to `at-ps2`
`modelf/6450225` keyboard has been renamed to a more suitable `modelf/pcat`

Keyboard descriptors for both Model F and Model M keyboards have also been updated.