Allow in-built mice to be overriden
Some keyboards which may have an inbuilt mouse, previously we would simply force this to be used and not allow the mouse to be overriden.

This allows users to still specify the `MOUSE` flag during compilation, and this will in effect override the one which may be specified within the keyboard configuration.

It may be the inbuit one is either not functional or broke, or it may be that it is preferred to use one external to the keyboard as-is.