Start addition of Apple M0110(A) Support.
Added experimental code to start work on supporting the Apple M0110(A) series of Keyboards.

This is purely hypothetical at present and has not been tested.  The following resources were used to piece together assumptions on how this may work:

https://www.apple.asimov.net/documentation/macintosh/Mac%20Hardware%20Info%20-%20Mac%20128K.pdf
https://www.apple.asimov.net/documentation/macintosh/Mac%20Hardware%20Info%20-%20Mac%20Plus.pdf

Page 20 from this resource for Protocol information
http://www.mac.linux-m68k.org/devel/plushw.php.html

I threw this together (so far) while on the bus home from work...